refactor: moved main server logic to a seperate module
